reorganized code
made most compoent files TitleCase. split bullets.js into Bullet.js and BulletComparator.js. did some other minor code cleanup
